pyproject.toml
[tool]
[tool.poetry]
name = "EyeTrackVR"
version = "1.0"
description = "Opensource, affordable VR eye tracker for VRChat"
authors = ["RedHawk989 <48768484+RedHawk989@users.noreply.github.com>"]
license = "GPL v3"
repository = "https://github.com/RedHawk989/ETVR-Data-Collection"

[tool.poetry.dependencies]
python = ">=3.10.0"
python-osc = "^1.8.1"
opencv-python = "^4.8.0.74"
numpy = "^1.23.5"
watchdog = "^3.0.0"
pyserial = "^3.5"
psutil = "^5.9.7"
colorama = "^0.4.6"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.21.1"
pyinstaller = "^5.6.2"
viztracer = "^0.15.6"
black = "^22.10.0"
pytest = "^7.2.0"
mypy = "^1.4.1"
ruff = "^0.1.5"

[tool.black]
line-length = 135
exclude = "(.git|.env|venv|.venv|build|dist|.vscode|.idea|__pycache__|.ruff_cache)"
target-version = ["py310", "py311"]

[tool.ruff]
select = ["E", "F", "W", "Q"]
src = ["TrackingBackend", "test"]
respect-gitignore = true
target-version = "py311"
output-format = "grouped"
indent-width = 4
exclude = ["__pycache__", ".ruff_cache", ".vscode", ".idea", ".venv", "build", "dist", ".git", ".env", "venv"]
line-length = 135

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
